+++ config.js
@@ -1,5 +1,0 @@
-window.SPOTIFY_CONFIG = {
-    clientId: 'YOUR_CLIENT_ID_HERE',
-    redirectUri: 'http://localhost:8080/',
-    syncUrl: null,
-};